refactor(backend): move fitness models to correct module
