+++ apps/backend/src/fitness/mod.rs
@@ -1,2 +1,3 @@
-pub mod logic;
 pub mod resolver;
+
+mod logic;
