+++ apps/backend/src/models.rs
@@ -1137,4 +1137,43 @@ pub mod fitness {
         pub lifetime_stats: TotalMeasurement,
         pub personal_bests: Vec<UserToExerciseBestSetExtraInformation>,
     }
+
+    #[derive(
+        Clone, Debug, Deserialize, Serialize, FromJsonQueryResult, Eq, PartialEq, SimpleObject,
+    )]
+    pub struct ProcessedExercise {
+        pub exercise_name: String,
+        pub exercise_id: i32,
+        pub sets: Vec<WorkoutSetRecord>,
+        pub notes: Vec<String>,
+        pub rest_time: Option<u16>,
+        pub total: TotalMeasurement,
+    }
+
+    #[derive(
+        Clone, Debug, Deserialize, Serialize, FromJsonQueryResult, Eq, PartialEq, SimpleObject,
+    )]
+    pub struct WorkoutInformation {
+        /// Each grouped superset of exercises will be in a vector. They will contain
+        /// the `exercise.idx`.
+        pub supersets: Vec<Vec<u16>>,
+        pub exercises: Vec<ProcessedExercise>,
+    }
+
+    #[derive(
+        Clone, Debug, Deserialize, Serialize, FromJsonQueryResult, Eq, PartialEq, SimpleObject,
+    )]
+    pub struct WorkoutSummaryExercise {
+        pub num_sets: usize,
+        pub name: String,
+        pub best_set: WorkoutSetRecord,
+    }
+
+    #[derive(
+        Clone, Debug, Deserialize, Serialize, FromJsonQueryResult, Eq, PartialEq, SimpleObject,
+    )]
+    pub struct WorkoutSummary {
+        pub total: TotalMeasurement,
+        pub exercises: Vec<WorkoutSummaryExercise>,
+    }
 }
